Moved third party header files from include to lib folder
