+++ test/src/subnet/main.cpp
@@ -14,8 +14,8 @@
 #include <cadmium/logger/common_loggers.hpp>
 
 
-#include "../../../include/NDTime.hpp"
-#include "../../../include/iestream.hpp"
+#include "../../../lib/NDTime.hpp"
+#include "../../../lib/iestream.hpp"
 
 #include "../../../include/message.hpp"
